Move example script to example/tune/, so it can be imported as module.
